# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Serwish Api",
        "description": "Serwish description",
        "contact": {
            "email": "dev@serwish.ge"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://serwish.ge",
            "description": "serwish web-site"
        },
        {
            "url": "https://api.serwish.ge/api",
            "description": "serwish api"
        }
    ],
    "paths": {
...
        "/api/posts/list": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get posts ",
                "operationId": "getFilteredPosts",
                "parameters": [
                    {
                        "name": "filterable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/PostListRequest"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "fail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{id}/{locale}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get posts ",
                "operationId": "getSinglePost",
                "responses": {
                    "404": {
                        "description": "fail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiRequestException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            },
...
            "PostListRequest": {
                "properties": {
                    "locale": {
                        "description": "locale",
                        "format": "locale",
                        "default": "ka"
                    },
                    "page": {
                        "description": "page",
                        "format": "integer",
                        "default": 1
                    },
                    "perPage": {
                        "description": "perPage",
                        "format": "integer",
                        "default": 20
                    },
                    "categories": {
                        "description": "categories you can specify multiple ids",
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "type": "string",
                                    "default": "22,33,etc",
                                    "property": "categories"
                                }
                            ]
                        }
                    },
                    "filterBy": {
                        "description": "Get the validation rules that apply to the request.",
                        "type": "enum",
                        "enum": [
                            "monthly",
                            "weekly",
                            "newest"
                        ]
                    }
                },
                "type": "object"
            },
...
            "Post": {
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "format": "int64"
                    },
                    "author_id": {
                        "title": "author_id",
                        "description": "author_id",
                        "format": "int64"
                    },
                    "viewCount": {
                        "title": "viewCount",
                        "description": "viewCount",
                        "format": "int64"
                    },
                    "isActive": {
                        "title": "isActive",
                        "description": "isActive",
                        "format": "boolean"
                    },
                    "publish_at": {
                        "title": "publish_at",
                        "description": "publish_at",
                        "format": "string"
                    },
                    "translated": {
                        "$ref": "#/components/schemas/PostTranslation"
                    }
                },
                "type": "object"
            },
            "PostTranslation": {
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "format": "int64"
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "format": "string"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "format": "string"
                    },
                    "slug": {
                        "title": "slug",
                        "description": "slug",
                        "format": "string"
                    },
                    "meta": {
                        "title": "meta",
                        "description": "meta",
                        "format": "string"
                    },
                    "locale": {
                        "title": "locale",
                        "description": "locale",
                        "format": "string"
                    },
                    "post_id": {
                        "title": "post_id",
                        "description": "post_id",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global Authorization: Bearer token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/token",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token",
                        "scopes": {}
                    }
                },
                "scheme": "https"
            }
        }
    },
    "security": [
        []
    ]
}